e2e: Improve `WaitForSRIOVStable()`
When the `WaitForSRIOVStable()` function fails, it outputs something like:
```
{Timed out after 1200.000s.
Expected
    <bool>: false
to be true failed [FAILED] Timed out after 1200.000s.
Expected
    <bool>: false
to be true
In [It] at: /tmp/cnf-gaH5c/cnf-features-deploy/cnf-tests/submodules/sriov-network-operator/test/conformance/tests/test_sriov_operator.go:2087 @ 04/04/25 23:07:39.673
}
```

Which is not helpful to understand the root cause.

Add information about the SriovNetworkNodeStates and the latest events in the
operator's namespace

Signed-off-by: Andrea Panattoni <apanatto@redhat.com>