Remove unused files from `/config`
Folder `/config` contains the information to build an operator-sdk bundle, which logic
has been removed by #502.
They progressively diverged from the openshift/sriov-network-operator fork, and the files
in these repository not being used became obsolete, with less information or wrong data at all.

As a first step to clean, remove all the unused files in the folder. `config/crd/bases` is still
a good source of truth, as it is used by the `manifests` Makefile rule to update the Helm charts.

In the future, if we want to support OLM deployment in this repository, we can
renovate the configuration starting from the `openshift` fork, maybe adjusting
the CI to validate its content at every PR.

Refs
- https://github.com/k8snetworkplumbingwg/sriov-network-operator/pull/502

Signed-off-by: Andrea Panattoni <apanatto@redhat.com>